+++ collective/computedfield/utils.py
@@ -1,4 +1,9 @@
+import calendar
+import datetime
 import json
+import time
+
+import pytz
 
 try:
     import persistent
@@ -21,8 +26,67 @@ is_pmap = lambda v: HAS_ZODB and isinstance(v, PersistentMapping)
 is_btree = lambda v: 'BTree' in v.__class__.__name__
 
 
+def days_since_epoch(value):
+    if not isinstance(value, datetime.datetime):
+        dt = datetime.datetime(
+            value.year,
+            value.month,
+            value.day,
+            tzinfo=pytz.UTC,
+            )
+    else:
+        dt = value
+    if dt.tzinfo is None:
+        # treat naive as UTC by reconstructing as much:
+        dt = datetime.datetime(
+            value.year,
+            value.month,
+            value.day,
+            value.hour,
+            value.minute,
+            value.second,
+            value.microsecond,
+            tzinfo=pytz.UTC,
+            )
+    return calendar.timegm(dt.utctimetuple()) / 86400.0
+
+
+def parse_datestamp(value):
+    y = int(value[0:4])
+    m = int(value[5:7])
+    d = int(value[8:10])
+    return datetime.datetime(y, m, d, tzinfo=pytz.UTC)
+
+
+def normalize_value(value):
+    """
+    Normalize non-float value to floating-point.  Supports:
+
+        * int -> float
+        * Strings that look like numbers
+        * ISO 8601 date stamps (not time components)
+        * datetime.date objects
+        * datetime.datetime (naive as UTC, or pytz.UTC) objects
+    """
+    if isinstance(value, int):
+        value = float(value)
+    if isinstance(value, basestring):
+        try:
+            # is it string that looks like a number?
+            value = float(value)
+        except ValueError:
+            # if it looks like a datestamp, parse it:
+            if len(value) == 10 and value.find('-') == 4:
+                value = parse_datestamp(value)
+    if any(
+        [isinstance(value, cls) for cls in (datetime.date, datetime.datetime)]
+            ):
+        value = days_since_epoch(value)
+    return value
+
+
 def normalize_data(context):
-    """Return a dict of context data"""
+    """Return a dict of context data; does not normalize values"""
     if isinstance(context, basestring):
         context = json.loads(context)
     if isinstance(context, dict):
@@ -33,6 +97,7 @@ def normalize_data(context):
             return dict(data)
         # unghost, if necessary to get __dict__ on object:
         v = getattr(data, 'BOGUS_ATTR_NAME', None)  # noqa
-    return getattr(context, '__dict__', {})  # attrs as mapping
-
+    else:
+        data = context
+    return _dict(data)  # attrs as mapping